+++ src/main/java/com/logicartisan/common/core/listeners/package-info.java
@@ -1,0 +1,7 @@
+/**
+ *
+ */
+@ParametersAreNonnullByDefault
+package com.logicartisan.common.core.listeners;
+
+import javax.annotation.ParametersAreNonnullByDefault;